+++ Util/Assets.cs
@@ -14,9 +14,11 @@ namespace FacilityMeltdown {
         internal static AssetBundle assetBundle;
 
         internal static AudioClip[] warnings;
-        internal static AudioClip music, shockwave;
-        internal static GameObject facilityExplosionPrefab, meltdownHandlerPrefab, shockwavePrefab;
+        internal static AudioClip music, shockwave, scannerStart;
+        internal static GameObject facilityExplosionPrefab, meltdownHandlerPrefab, shockwavePrefab, geigerCounterItem;
         internal static GameObject[] facilityEffects;
+        internal static Item geigerCounterItemDef;
+        internal static TerminalNode geigerCounterNode;
         public static void Init() {
             assetBundle = AssetBundle.LoadFromFile(Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "facilitymeltdown"));
 
@@ -29,10 +31,15 @@ namespace FacilityMeltdown {
 
             music = assetBundle.LoadAsset<AudioClip>("music.mp3");
             shockwave = assetBundle.LoadAsset<AudioClip>("shockwave.mp3");
+            scannerStart = assetBundle.LoadAsset<AudioClip>("scannerStart.mp3");
 
             facilityExplosionPrefab = assetBundle.LoadAsset<GameObject>("FacilityExplosion.prefab");
             meltdownHandlerPrefab = assetBundle.LoadAsset<GameObject>("MeltdownHandler.prefab");
             shockwavePrefab = assetBundle.LoadAsset<GameObject>("Shockwave.prefab");
+
+            geigerCounterItem = assetBundle.LoadAsset<GameObject>("GeigerCounterItem.prefab");
+            geigerCounterItemDef = assetBundle.LoadAsset<Item>("GeigerCounterItemDef.asset");
+            geigerCounterNode = assetBundle.LoadAsset<TerminalNode>("GeigerCounterNode.asset");
 
             facilityEffects = new GameObject[] {
                 assetBundle.LoadAsset<GameObject>("Dust.prefab"),